requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.1
colorlog==6.9.0
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==5.2
django-appconf==1.1.0
django-ckeditor==6.7.2
django-classy-tags==4.1.0
django-constance==4.3.2
django-crispy-forms==2.3
django-filter==24.3
django-js-asset==3.1.2
django-sekizai==4.1.0
django-select2==8.4.0
django-sendfile2==0.7.1
django-sitetree==1.18.0
django-tables2==2.7.5
django-taggit==6.1.0
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
frozenlist==1.5.0
geoip2==4.8.1
hashids==1.3.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
lxml==5.3.1
maxminddb==2.6.3
multidict==6.2.0
pillow==11.1.0
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
pyOpenSSL==25.0.0
requests==2.32.3
service-identity==24.2.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn
psycopg
celery[redis,auth,gevent]
flower
django-celery-results
django-celery-beat
